% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{add_created_by}
\alias{add_created_by}
\title{Add createdBy key to metdata}
\usage{
add_created_by(d)
}
\arguments{
\item{list}{d: Metadata}
}
\value{
list d: Metadata
}
\description{
Add createdBy key to metdata
}
\keyword{internal}

