% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_merge.R
\name{add_csv_to_columns}
\alias{add_csv_to_columns}
\title{Add csv data to each column in a list of columns}
\usage{
add_csv_to_columns(csv.cols, meta.cols)
}
\arguments{
\item{list}{csv.cols: Values, sorted by column}

\item{list}{meta.cols: Metadata, sorted by column}
}
\value{
list meta.cols: Metadata
}
\description{
Add csv data to each column in a list of columns
}
\keyword{internal}

