% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_merge_data.R
\name{addCsvToMetadataLoad}
\alias{addCsvToMetadataLoad}
\title{Merge Main. Call the individual steps of merge for each file.}
\usage{
addCsvToMetadataLoad(d)
}
\arguments{
\item{d}{LiPD Metdata}
}
\value{
d Modfified LiPD Metadata
}
\description{
Merge Main. Call the individual steps of merge for each file.
}
\keyword{internal}

