% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_merge_data.R
\name{addCsvToMetadataRead}
\alias{addCsvToMetadataRead}
\title{Merge Main. Call the individual steps of merge for each file.}
\usage{
addCsvToMetadataRead(d)
}
\arguments{
\item{d}{LiPD Metdata}
}
\value{
d Modfified LiPD Metadata
}
\description{
Merge Main. Call the individual steps of merge for each file.
}
\keyword{internal}

