% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{ask_how_many}
\alias{ask_how_many}
\title{Ask if user wants to read one file or a directory with multiple files.}
\usage{
ask_how_many()
}
\value{
ans Answer to prompt (s/m)
}
\description{
Ask if user wants to read one file or a directory with multiple files.
}
\keyword{internal}

