% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_helpers.R
\name{askHowMany}
\alias{askHowMany}
\title{Ask if user wants to load one file or a directory with multiple files.}
\usage{
askHowMany()
}
\value{
ans Answer to prompt (s/m)
}
\description{
Ask if user wants to load one file or a directory with multiple files.
}
\keyword{internal}

