% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{browse_dialog}
\alias{browse_dialog}
\title{Open a file browsing gui to let the user pick a file or directory}
\usage{
browse_dialog(ans)
}
\arguments{
\item{char}{ans single (s) or directory (d)}
}
\value{
char path Directory or file path
}
\description{
Open a file browsing gui to let the user pick a file or directory
}
\keyword{internal}

