% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{browseDialog}
\alias{browseDialog}
\title{Open a file browsing gui to let the user pick a location}
\usage{
browseDialog(ans)
}
\arguments{
\item{ans}{Single or multiple files}
}
\value{
path Path to file
}
\description{
Open a file browsing gui to let the user pick a location
}
\keyword{internal}

