% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{build_structure}
\alias{build_structure}
\title{Before you place a table, you must have the structure leading up to the location or you'll get errors. Build the structure}
\usage{
build_structure(d, pc, table_type, pcNumber, modelNumber, tableNumber)
}
\arguments{
\item{list}{d: Metadata}

\item{char}{table_type: meas, ens, or summ}

\item{num}{tableumber: Table index number}
}
\value{
list d: Metadata
}
\description{
Before you place a table, you must have the structure leading up to the location or you'll get errors. Build the structure
}
