% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_versions.R
\name{checkVersion}
\alias{checkVersion}
\title{Check the version number from metadata}
\usage{
checkVersion(d)
}
\arguments{
\item{d}{LiPD Metadata}
}
\value{
version LiPD version number
}
\description{
Check the version number from metadata
}
\keyword{internal}

