% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{collapse_column}
\alias{collapse_column}
\title{Collapse time series column. Compile column entries and place new column in table.}
\usage{
collapse_column(table, entry, pc)
}
\arguments{
\item{list}{table: Metadata}

\item{list}{entry: Time series entry}

\item{char}{pc: paleoData or chronData}
}
\value{
list table: Metadata
}
\description{
Collapse time series column. Compile column entries and place new column in table.
}
