% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{collapseTs}
\alias{collapseTs}
\title{Collapse time series into LiPD dataset form}
\usage{
collapseTs(ts)
}
\arguments{
\item{ts}{Time series}
}
\value{
D LiPD data, sorted by dataset name
}
\description{
Collapse time series into LiPD dataset form
}

