% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_split_csv.R, R/write_lipds_csv.R
\name{collectCsvSection}
\alias{collectCsvSection}
\title{Collect and remove csv from one section: Paleo or chron
csv.data format: [ some_filename.csv $columns.data ]}
\usage{
collectCsvSection(all.data, pc, dsn)

collectCsvSection(all.data, pc, dsn)
}
\arguments{
\item{pc}{paleoData or chronData}

\item{dsn}{Dataset name}

\item{d}{Metadata w. values}

\item{keys}{Section keys}

\item{csv.data}{Running collection of csv data}

\item{name}{LiPD data set name}

\item{d}{Metadata w. values}

\item{csv.data}{Running collection of csv data}
}
\value{
all.data List holding the running collection of separated csv and metadata

all.data List holding the running collection of separated csv and metadata
}
\description{
Collect and remove csv from one section: Paleo or chron
csv.data format: [ some_filename.csv $columns.data ]

Collect and remove csv from one section: Paleo or chron
csv.data format: [ some_filedsn.csv $columns.data ]
}
\keyword{internal}

