% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_helpers.R
\name{colsToList}
\alias{colsToList}
\title{Convert column type: data frame to list
Function may be unused right now}
\usage{
colsToList(table)
}
\arguments{
\item{table}{Data table}
}
\value{
table Converted data table
}
\description{
Convert column type: data frame to list
Function may be unused right now
}
\keyword{internal}

