% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_extract.R
\name{combineInterpretationByScope}
\alias{combineInterpretationByScope}
\title{Combine interpretations by scope}
\usage{
combineInterpretationByScope(sTS)
}
\arguments{
\item{sTS}{}
}
\value{
a regular Timeseries structure
}
\description{
Combine interpretations by scope
}
