% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{confirmLipdVersion}
\alias{confirmLipdVersion}
\title{Check for a LiPD version in the metadata. Create one if it isn't found.}
\usage{
confirmLipdVersion(j)
}
\arguments{
\item{j}{Json metadata}
}
\value{
j Json metadata
}
\description{
Check for a LiPD version in the metadata. Create one if it isn't found.
}
\keyword{internal}

