% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_versions.R
\name{convertDfsLst}
\alias{convertDfsLst}
\title{Check / convert and fixed data frames into scalable lists}
\usage{
convertDfsLst(d)
}
\arguments{
\item{d}{LiPD metadata}
}
\value{
d Modified LiPD metadata
}
\description{
Check / convert and fixed data frames into scalable lists
}
\keyword{internal}

