% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_versions.R
\name{convertSM}
\alias{convertSM}
\title{Convert from a single fixed table, to a multiple scalable table
(LiPD Verison 1.1 to 1.2 change)}
\usage{
convertSM(d, keys)
}
\arguments{
\item{d}{LiPD metadata}
}
\value{
d Modified LiPD metadata
}
\description{
Convert from a single fixed table, to a multiple scalable table
(LiPD Verison 1.1 to 1.2 change)
}
\keyword{internal}

