% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_range}
\alias{create_range}
\title{Create the range for ensemble table "number" field}
\usage{
create_range(start, len)
}
\arguments{
\item{start}{Number to start at}

\item{len}{Amount of times to loop}
}
\value{
l A vectory of column ints
}
\description{
Create the range for ensemble table "number" field
}
\keyword{internal}

