% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{create_tmp_dir}
\alias{create_tmp_dir}
\title{'Create' a temp directory (really its the same directory path for the whole R session, so it's not that new)
Recreate it if it's non-existant}
\usage{
create_tmp_dir()
}
\value{
char path:
}
\description{
'Create' a temp directory (really its the same directory path for the whole R session, so it's not that new)
Recreate it if it's non-existant
}
\keyword{internal}
