% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{createTmpDir}
\alias{createTmpDir}
\title{Create a temporary working directory}
\usage{
createTmpDir()
}
\value{
d Temporary directory path
}
\description{
Create a temporary working directory
}
\keyword{internal}

