% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_num.R
\name{export_data}
\alias{export_data}
\title{Index a section. paleoData or chronData}
\usage{
export_data(section, pc)
}
\arguments{
\item{list}{section: Metadata}

\item{char}{pc: paleoData or chronData}
}
\value{
list d: Metadata
}
\description{
Index a section. paleoData or chronData
}
\keyword{internal}

