% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_extract.R
\name{extractTs}
\alias{extractTs}
\title{Extract time series from LiPD data}
\usage{
extractTs(D, chron = NULL)
}
\arguments{
\item{D}{LiPD data, sorted by dataset name}
}
\value{
ts Time series
}
\description{
Extract time series from LiPD data
}

