% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag.R
\name{fake_bagit}
\alias{fake_bagit}
\title{Run a fake version of bagit. No bagit txt files are created}
\usage{
fake_bagit(dir_bag)
}
\arguments{
\item{char}{dir_bag: The path to the directory that needs to be bagged}
}
\value{
none
}
\description{
Run a fake version of bagit. No bagit txt files are created
}
\keyword{internal}
