% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{get_crumbs}
\alias{get_crumbs}
\title{Use regex to split the tableName into a crumbs path}
\usage{
get_crumbs(ts)
}
\arguments{
\item{list}{ts: Time series}
}
\value{
list matches: Separated names and indices
}
\description{
Use regex to split the tableName into a crumbs path
}

