% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_split_csv.R
\name{get_csv_from_columns}
\alias{get_csv_from_columns}
\title{Parse metadata and csv from columns}
\usage{
get_csv_from_columns(table)
}
\arguments{
\item{list}{table: Metadata}
}
\value{
list new: Metadata / CSV
}
\description{
Parse metadata and csv from columns
}
\keyword{internal}

