% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_split_csv.R
\name{get_csv_from_json}
\alias{get_csv_from_json}
\title{Get csv "values" fields from metadata.}
\usage{
get_csv_from_json(name, d)
}
\arguments{
\item{char}{name: Name of current LiPD record}

\item{list}{d: Metadata}
}
\value{
all.data Final split of metadata and csv data
}
\description{
Get csv "values" fields from metadata.
}
\keyword{internal}

