% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_split_csv.R
\name{get_csv_from_metadata}
\alias{get_csv_from_metadata}
\title{Get csv "values" fields from metadata.}
\usage{
get_csv_from_metadata(d, dsn)
}
\arguments{
\item{list}{d: Metadata}

\item{char}{dsn: Dataset name}
}
\value{
list d2: Metadata / CSV
}
\description{
Get csv "values" fields from metadata.
}
\keyword{internal}
