% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_split_csv.R
\name{get_csv_from_model}
\alias{get_csv_from_model}
\title{Parse metadata and csv from models}
\usage{
get_csv_from_model(models, crumbs, csvs)
}
\arguments{
\item{list}{models: Metadata}

\item{char}{crumbs: Crumbs}

\item{list}{csvs: CSV data}
}
\value{
list new: Metadata / CSV
}
\description{
Parse metadata and csv from models
}
\keyword{internal}
