% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_split_csv.R
\name{get_csv_from_section}
\alias{get_csv_from_section}
\title{Get CSV from one section
csv.data format: [ some_filename.csv $columns.data ]}
\usage{
get_csv_from_section(dat, pc_1, dsn)
}
\arguments{
\item{list}{dat: Metadata}

\item{char}{pc_1: paleo or chron}

\item{char}{dsn: Dataset name}
}
\value{
list dat: Split data
}
\description{
Get CSV from one section
csv.data format: [ some_filename.csv $columns.data ]
}
\keyword{internal}
