% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_split_csv.R
\name{get_csv_from_table}
\alias{get_csv_from_table}
\title{Parse metadata and csv from list of tables}
\usage{
get_csv_from_table(tables, crumbs, csvs)
}
\arguments{
\item{list}{tables: Metadata}

\item{char}{crumbs: Crumbs}

\item{list}{csvs: CSV data}
}
\value{
list new: Metadata / CSV
}
\description{
Parse metadata and csv from list of tables
}
\keyword{internal}
