% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_datasetname}
\alias{get_datasetname}
\title{Get dataSetName from metadata. If one is not found, use filename as fallback.}
\usage{
get_datasetname(d, name)
}
\arguments{
\item{list}{d: Metadata}

\item{char}{name: Filename fallback}
}
\value{
char dsn: Dataset name
}
\description{
Get dataSetName from metadata. If one is not found, use filename as fallback.
}
\keyword{internal}
