% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_download_path}
\alias{get_download_path}
\title{Locate a folder to download a file to}
\usage{
get_download_path()
}
\value{
char dst_path: Destination directory
}
\description{
Locate a folder to download a file to
}
\keyword{internal}
