% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_lipd_paths}
\alias{get_lipd_paths}
\title{Get a list of paths to LiPD files}
\usage{
get_lipd_paths(path)
}
\arguments{
\item{path}{Directory or file path}
}
\value{
list files File paths to LiPD files
}
\description{
Get a list of paths to LiPD files
}
\keyword{internal}

