% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{get_lipd_version}
\alias{get_lipd_version}
\title{Check what version of LiPD this file is using. If none is found, assume it's using version 1.0}
\usage{
get_lipd_version(d)
}
\arguments{
\item{list}{d: Metadata}
}
\value{
numeric version: Version number
}
\description{
Check what version of LiPD this file is using. If none is found, assume it's using version 1.0
}
\keyword{internal}

