% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_src_or_dst}
\alias{get_src_or_dst}
\title{Ask user where local file/directory location is.}
\usage{
get_src_or_dst(path)
}
\arguments{
\item{char}{path Target path}
}
\value{
char path Directory or file path
}
\description{
Ask user where local file/directory location is.
}
\keyword{internal}

