% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{get_ts_lipd}
\alias{get_ts_lipd}
\title{Retreieve the original datasets from the lipd R environment}
\usage{
get_ts_lipd()
}
\value{
list tmp_storage: Temporary storage data that holds original datasets
}
\description{
Retreieve the original datasets from the lipd R environment
}
