% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_file.R
\name{getDataRead}
\alias{getDataRead}
\title{Retrieve and import csv and jsonld files in the current directory.}
\usage{
getDataRead()
}
\value{
data.list List of data for one LiPD file
}
\description{
Retrieve and import csv and jsonld files in the current directory.
}
\keyword{internal}

