% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_main.R
\name{getDatasetName}
\alias{getDatasetName}
\title{Get data set name from metadata}
\usage{
getDatasetName(d)
}
\arguments{
\item{d}{LiPD data}
}
\value{
name Data set name
}
\description{
Get data set name from metadata
}
\keyword{internal}

