% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_helpers.R
\name{getListLpdExt}
\alias{getListLpdExt}
\title{#' Import each of the required modules for the package
#' @export
#' @keywords internal
#' @return none
setModules <- function(){
  suppressPackageStartupMessages(library(tools, quietly=TRUE, verbose=FALSE))
  suppressPackageStartupMessages(library(Kmisc, quietly=TRUE, verbose=FALSE))
  suppressPackageStartupMessages(library(RJSONIO, quietly=TRUE, verbose=FALSE))
  suppressPackageStartupMessages(library(jsonlite, quietly=TRUE, verbose=FALSE))
  suppressPackageStartupMessages(library(BBmisc, quietly=TRUE, verbose=FALSE))
}
Get list of all LiPD files in current directory}
\usage{
getListLpdExt(path.and.file)
}
\arguments{
\item{path.and.file}{Target directory and 1+ files}
}
\value{
f List of LiPD files w. ext
}
\description{
#' Import each of the required modules for the package
#' @export
#' @keywords internal
#' @return none
setModules <- function(){
  suppressPackageStartupMessages(library(tools, quietly=TRUE, verbose=FALSE))
  suppressPackageStartupMessages(library(Kmisc, quietly=TRUE, verbose=FALSE))
  suppressPackageStartupMessages(library(RJSONIO, quietly=TRUE, verbose=FALSE))
  suppressPackageStartupMessages(library(jsonlite, quietly=TRUE, verbose=FALSE))
  suppressPackageStartupMessages(library(BBmisc, quietly=TRUE, verbose=FALSE))
}
Get list of all LiPD files in current directory
}
\keyword{internal}

