% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvs.R
\name{getMeasurementTables}
\alias{getMeasurementTables}
\title{Get measTables}
\usage{
getMeasurementTables(L, pc = "all")
}
\arguments{
\item{L}{a Lipd file}

\item{pc}{paleo or chron tables? (default= "all)}
}
\value{
a list of data.frames
}
\description{
Get measTables
}
