% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getSrcOrDst}
\alias{getSrcOrDst}
\title{Ask user where local file/folder location is.}
\usage{
getSrcOrDst()
}
\value{
path.and.file Path to files
}
\description{
Ask user where local file/folder location is.
}
\keyword{internal}

