% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_num.R
\name{idx_col_by_num}
\alias{idx_col_by_num}
\title{Remove column names indexing. Set them to index by their column number
Place the new columns under a "columns" list}
\usage{
idx_col_by_num(table)
}
\arguments{
\item{list}{table: Metadata}
}
\value{
list table: Metadata
}
\description{
Remove column names indexing. Set them to index by their column number
Place the new columns under a "columns" list
}
\keyword{internal}
