% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_num.R
\name{idx_name_to_num}
\alias{idx_name_to_num}
\title{Switch all indexing from names to numbers.}
\usage{
idx_name_to_num(d)
}
\arguments{
\item{list}{d: Metadata}
}
\value{
list d: Metadata
}
\description{
Switch all indexing from names to numbers.
}
\keyword{internal}
