% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_name.R
\name{idx_num_to_name}
\alias{idx_num_to_name}
\title{Change index-by-number to index-by-variableName}
\usage{
idx_num_to_name(d)
}
\arguments{
\item{list}{d: Metadata}
}
\value{
list d: Metadata
}
\description{
Change index-by-number to index-by-variableName
}
\keyword{internal}

