% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_name.R
\name{idx_table_by_name}
\alias{idx_table_by_name}
\title{Index tables in a loop}
\usage{
idx_table_by_name(tables)
}
\arguments{
\item{list}{tables: Metadata}
}
\value{
list tables: Metadata
}
\description{
Index tables in a loop
}
\keyword{internal}
