% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_num.R
\name{idx_table_by_num_what}
\alias{idx_table_by_num_what}
\title{Unindex tables by name. Move to index by number.}
\usage{
idx_table_by_num_what(table, pc, tableType)
}
\arguments{
\item{table}{Table data}
}
\value{
table Modified table data
}
\description{
Unindex tables by name. Move to index by number.
}
\keyword{internal}

