% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{index_geo}
\alias{index_geo}
\title{Make geo semi-flat. Remove unnecessary levels between us and data.}
\usage{
index_geo(d)
}
\arguments{
\item{d}{Metadata}
}
\value{
d Modified metadata
}
\description{
Make geo semi-flat. Remove unnecessary levels between us and data.
}
\keyword{internal}
