% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_indexing.R
\name{indexByNameRead}
\alias{indexByNameRead}
\title{Change index-by-number to index-by-variableName}
\usage{
indexByNameRead(d)
}
\arguments{
\item{d}{LiPD file}
}
\value{
d Modified LiPD file
}
\description{
Change index-by-number to index-by-variableName
}
\keyword{internal}

