% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_indexing.R
\name{indexByNumberSave}
\alias{indexByNumberSave}
\title{Main indexing. Convert all index-by-name to index-by-number.}
\usage{
indexByNumberSave(d)
}
\arguments{
\item{d}{Metadata}
}
\value{
d Modified metadata
}
\description{
Main indexing. Convert all index-by-name to index-by-number.
}
\keyword{internal}

