% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_indexing.R
\name{indexSectionLoad}
\alias{indexSectionLoad}
\title{Change index-by-number for one section}
\usage{
indexSectionLoad(d, keys)
}
\arguments{
\item{d}{LiPD metadata}

\item{keys}{Section keys}
}
\value{
d Modified LiPD metadata
}
\description{
Change index-by-number for one section
}
\keyword{internal}

