% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_indexing.R
\name{indexSectionSave}
\alias{indexSectionSave}
\title{Index a single section. Paleo or Chron}
\usage{
indexSectionSave(d, keys)
}
\arguments{
\item{d}{LiPD Metadata}

\item{keys}{Section keys}
}
\value{
d Modified metadata
}
\description{
Index a single section. Paleo or Chron
}
\keyword{internal}

