% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lipds_indexing.R
\name{indexSectionWrite}
\alias{indexSectionWrite}
\title{Index a single section. Paleo or Chron}
\usage{
indexSectionWrite(d, keys)
}
\arguments{
\item{d}{LiPD Metadata}

\item{keys}{Section keys}
}
\value{
d Modified metadata
}
\description{
Index a single section. Paleo or Chron
}
\keyword{internal}

