% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{isDirectory}
\alias{isDirectory}
\title{Checks if a path is a directory or not a directory}
\usage{
isDirectory(s)
}
\arguments{
\item{s}{Target path}
}
\value{
boolean
}
\description{
Checks if a path is a directory or not a directory
}
\keyword{internal}
