% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isNullOb}
\alias{isNullOb}
\title{Checks if an object is null/empty}
\usage{
isNullOb(x)
}
\arguments{
\item{x}{Data object to check}
}
\value{
boolean
}
\description{
Checks if an object is null/empty
}
\keyword{internal}
