% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{list_files_recursive}
\alias{list_files_recursive}
\title{Recursive file list for current directory and below}
\usage{
list_files_recursive(x)
}
\arguments{
\item{char}{x: File type}
}
\value{
char files: Matching file paths
}
\description{
Recursive file list for current directory and below
}
\keyword{internal}
