% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_helpers.R
\name{listFiles}
\alias{listFiles}
\title{Get list of csv files in current directory and below}
\usage{
listFiles(x)
}
\value{
f List of csv files
}
\description{
Get list of csv files in current directory and below
}
\keyword{internal}

