% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_main.R
\name{loadLipds}
\alias{loadLipds}
\title{Main LiPD loading function. Combines all processes into one.}
\usage{
loadLipds()
}
\value{
D LiPD Library
}
\description{
Main LiPD loading function. Combines all processes into one.
}
\keyword{internal}

