% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeTs.R
\name{lumpTsMetaVariables}
\alias{lumpTsMetaVariables}
\title{Lump metadata variables with other rows in Ts (treeTS)}
\usage{
lumpTsMetaVariables(Ts, sc = c("eps", "rbar", "ncores"))
}
\arguments{
\item{Ts}{}

\item{sc}{}
}
\value{
A lumped Ts
}
\description{
Lump metadata variables with other rows in Ts (treeTS)
}
