% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_merge.R
\name{merge_csv_columns}
\alias{merge_csv_columns}
\title{Merge values into each column}
\usage{
merge_csv_columns(csv.cols, meta.cols)
}
\arguments{
\item{list}{csv.cols: Values, sorted by column}

\item{list}{meta.cols: Metadata, sorted by column}
}
\value{
list meta.cols: Metadata
}
\description{
Merge values into each column
}
\keyword{internal}

