% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_merge.R
\name{merge_csv_section}
\alias{merge_csv_section}
\title{Merge csv numeric data into the metadata columns}
\usage{
merge_csv_section(section, pc, csv.data)
}
\arguments{
\item{list}{d: Metadata}

\item{char}{pc: paleoData or chronData}

\item{list}{csv.data: CSV, sorted by filename}
}
\value{
list d: Metadata
}
\description{
Merge csv numeric data into the metadata columns
}
\keyword{internal}

