% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_merge.R
\name{merge_csv_table}
\alias{merge_csv_table}
\title{Merge CSV data into each table}
\usage{
merge_csv_table(tables, crumbs, csvs)
}
\arguments{
\item{list}{models: Metadata}

\item{char}{crumbs: Crumbs}

\item{list}{csvs: CSV data}
}
\value{
list models: Metadata
}
\description{
Merge CSV data into each table
}
\keyword{internal}

