% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_merge_data.R
\name{mergeCsvLoad}
\alias{mergeCsvLoad}
\title{Merge CSV data into the metadata}
\usage{
mergeCsvLoad(csv.cols, meta.cols)
}
\arguments{
\item{csv.cols}{CSV data for this file}

\item{meta.cols}{Target metadata columns}
}
\value{
meta.cols Modified metadata columns
}
\description{
Merge CSV data into the metadata
}
\keyword{internal}

