% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_merge_data.R
\name{mergeDataRead}
\alias{mergeDataRead}
\title{Merge csv numeric data into the metadata columns}
\usage{
mergeDataRead(d, keys)
}
\arguments{
\item{d}{LiPD metadata}

\item{keys}{Paleo or Chron keys}
}
\value{
d Modified LiPD metadata
}
\description{
Merge csv numeric data into the metadata columns
}
\keyword{internal}

