% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_indexing.R
\name{moveColsDownSave}
\alias{moveColsDownSave}
\title{Remove column names indexing. Set them to index by their column number
Place the new columns under a "columns" list}
\usage{
moveColsDownSave(table)
}
\arguments{
\item{table}{Table data}
}
\value{
table Modified table data
}
\description{
Remove column names indexing. Set them to index by their column number
Place the new columns under a "columns" list
}
\keyword{internal}

