% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_name.R
\name{moveColsUp}
\alias{moveColsUp}
\title{Get rid of "columns" layer so that the columns data is directly beneath its corresponding table}
\usage{
moveColsUp(table)
}
\arguments{
\item{table}{Table to be reorganized}
}
\value{
table Modified table
}
\description{
Get rid of "columns" layer so that the columns data is directly beneath its corresponding table
}
\keyword{internal}

