% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_indexing.R
\name{moveTableUpSave}
\alias{moveTableUpSave}
\title{Unindex tables by name. Move to index by number.}
\usage{
moveTableUpSave(table, pc, tableType)
}
\arguments{
\item{table}{Table data}
}
\value{
table Modified table data
}
\description{
Unindex tables by name. Move to index by number.
}
\keyword{internal}

