% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_main.R
\name{multiLoad}
\alias{multiLoad}
\title{Loop over multiple LiPD files. Load one at a time.}
\usage{
multiLoad(lpds_ext, working.dir, tmp)
}
\value{
D LiPD Library
}
\description{
Loop over multiple LiPD files. Load one at a time.
}
\keyword{internal}

