% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_main.R
\name{multiRead}
\alias{multiRead}
\title{Loop over multiple LiPD files. Read one at a time.}
\usage{
multiRead(lpds_ext, working.dir, tmp)
}
\value{
D LiPD Library
}
\description{
Loop over multiple LiPD files. Read one at a time.
}
\keyword{internal}

