% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lipds_main.R
\name{multiWrite}
\alias{multiWrite}
\title{Parse library of data records}
\usage{
multiWrite(D)
}
\arguments{
\item{D}{LiPD data}
}
\value{
none
}
\description{
Parse library of data records
}
\keyword{internal}

