% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_split_csv.R, R/write_lipds_csv.R
\name{parseTable}
\alias{parseTable}
\title{Parse the csv value columns from the table, then split the metadata from the csv}
\usage{
parseTable(table)

parseTable(table)
}
\arguments{
\item{table}{Table of data}

\item{table}{Table of data}
}
\value{
table Table w/o csv, csv Value columns

table Table w/o csv, csv Value columns
}
\description{
Parse the csv value columns from the table, then split the metadata from the csv

Parse the csv value columns from the table, then split the metadata from the csv
}
\keyword{internal}

