% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lipds_main.R
\name{promptStringWrite}
\alias{promptStringWrite}
\title{Prompt the user for some string entry}
\usage{
promptStringWrite(statement)
}
\arguments{
\item{statement}{Statement that gets printed to the user}
}
\value{
ans User input entry
}
\description{
Prompt the user for some string entry
}
\keyword{internal}

