% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{put_table}
\alias{put_table}
\title{Put the target table}
\usage{
put_table(d, m, pc, table)
}
\arguments{
\item{list}{d: Metadata}

\item{list}{m: Regex match}

\item{char}{pc: paleoData or chronData}

\item{list}{table: Metadata (to be placed)}
}
\value{
list d: Metadata
}
\description{
Put the target table
}

