% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{queryTs}
\alias{queryTs}
\title{Find all the time series entries that match a given search expression, 
and return vector of the indices that match. 
indices = queryTs(ts, "archiveType == marine sediment")
Valid operators : ==, =, <=, >=, <, >}
\usage{
queryTs(ts, expression)
}
\arguments{
\item{ts}{Time series}

\item{expression}{Search expression}
}
\value{
idx Matching indices
}
\description{
Find all the time series entries that match a given search expression, 
and return vector of the indices that match. 
indices = queryTs(ts, "archiveType == marine sediment")
Valid operators : ==, =, <=, >=, <, >
}

