% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvs.R
\name{read_csv_from_file}
\alias{read_csv_from_file}
\title{Opens the target CSV file and creates a dictionary with one list for each CSV column.}
\usage{
read_csv_from_file()
}
\value{
data.list List of data for one LiPD file
}
\description{
Opens the target CSV file and creates a dictionary with one list for each CSV column.
}
\keyword{internal}

