% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons.R
\name{read_jsonld}
\alias{read_jsonld}
\title{Find jsonld file in the cwd (or within a 2 levels below cwd), and load it in.}
\usage{
read_jsonld()
}
\value{
data.list List of data for one LiPD file
}
\description{
Find jsonld file in the cwd (or within a 2 levels below cwd), and load it in.
}
\keyword{internal}

