% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{readLipd}
\alias{readLipd}
\title{Read LiPD files into R workspace}
\usage{
readLipd(path = NULL)
}
\arguments{
\item{path}{Source path (optional)}
}
\value{
D LiPD data
}
\description{
Read LiPD files into R workspace
}
\keyword{internal}

