% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_file.R
\name{readLipdFile}
\alias{readLipdFile}
\title{Import the data from each csv and jsonld file for given LiPD}
\usage{
readLipdFile(lpd_noext, tmp)
}
\arguments{
\item{lpd_noext}{List of lipd files without extention}

\item{tmp}{Char path to the temp folder in memory}
}
\value{
out.list List of data for each lipd file
}
\description{
Import the data from each csv and jsonld file for given LiPD
}
\keyword{internal}

