% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{removeEmptyRec}
\alias{removeEmptyRec}
\title{Remove all NA, NULL, and empty objects from the data structure}
\usage{
removeEmptyRec(x)
}
\arguments{
\item{x}{Data structure}
}
\value{
x Modified data structure
}
\description{
Remove all NA, NULL, and empty objects from the data structure
}
\keyword{internal}

