% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_helpers.R
\name{removeLayers}
\alias{removeLayers}
\title{Remove CSV and metadata layer from LiPD file. Also, remove empties}
\usage{
removeLayers(d)
}
\arguments{
\item{d}{LiPD file}
}
\value{
d Modified LiPD file
}
\description{
Remove CSV and metadata layer from LiPD file. Also, remove empties
}
\keyword{internal}

