% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{replace_invalid_chars}
\alias{replace_invalid_chars}
\title{Check if output datasetname has invalid characters. Replace if necessary
R will not zip directories with certain characters.}
\usage{
replace_invalid_chars(dsn)
}
\arguments{
\item{char}{dsn: Dataset name}
}
\value{
char dsn: Dataset name
}
\description{
Check if output datasetname has invalid characters. Replace if necessary
R will not zip directories with certain characters.
}
\keyword{internal}
