% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rm_empty_fields}
\alias{rm_empty_fields}
\title{Remove all NA, NULL, and empty objects from the data structure}
\usage{
rm_empty_fields(x)
}
\arguments{
\item{x}{Data structure}
}
\value{
x Modified data structure
}
\description{
Remove all NA, NULL, and empty objects from the data structure
}
\keyword{internal}
