% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{rm_existing_tables}
\alias{rm_existing_tables}
\title{Remove any measurement or summary tables in the metadata for the given mode}
\usage{
rm_existing_tables(d, pc)
}
\arguments{
\item{list}{d: Metadata}

\item{char}{pc: paleoData or chronData}
}
\value{
list d: Metadata
}
\description{
Remove any measurement or summary tables in the metadata for the given mode
}

