% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_file.R
\name{saveLipdFile}
\alias{saveLipdFile}
\title{Main function. Run all save sub-routines for one LiPD record}
\usage{
saveLipdFile(d, name)
}
\arguments{
\item{d}{Metadata}

\item{name}{Name of current LiPD record}
}
\value{
none
}
\description{
Main function. Run all save sub-routines for one LiPD record
}
\keyword{internal}

