% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_main.R
\name{saveLipds}
\alias{saveLipds}
\title{Main function. Loop through all LiPD records, saving, and then cleaning up}
\usage{
saveLipds(D)
}
\arguments{
\item{D}{LiPD Library}
}
\value{
none
}
\description{
Main function. Loop through all LiPD records, saving, and then cleaning up
}

