% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag.R
\name{set_bagit}
\alias{set_bagit}
\title{Look for the users bagit script file. If you can't find it, ask them for help and/or ask them 
to download it and locate the file}
\usage{
set_bagit()
}
\value{
none
}
\description{
Look for the users bagit script file. If you can't find it, ask them for help and/or ask them 
to download it and locate the file
}
\keyword{internal}
