% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_versions.R
\name{setVersion}
\alias{setVersion}
\title{Set the LiPD version field}
\usage{
setVersion(d, ver)
}
\arguments{
\item{d}{LiPD Metadata}

\item{ver}{Version number}
}
\value{
d Modified LiPD Metadata
}
\description{
Set the LiPD version field
}
\keyword{internal}

