% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_main.R
\name{singleLoad}
\alias{singleLoad}
\title{Load one LiPD file. All steps}
\usage{
singleLoad(lpd, working.dir, tmp)
}
\value{
d LiPD File
}
\description{
Load one LiPD file. All steps
}
\keyword{internal}

