% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_main.R
\name{singleRead}
\alias{singleRead}
\title{Read one LiPD file. All steps}
\usage{
singleRead(lpd, working.dir, tmp)
}
\value{
d LiPD File
}
\description{
Read one LiPD file. All steps
}
\keyword{internal}

