% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lipds_main.R
\name{singleWrite}
\alias{singleWrite}
\title{Parse one data record}
\usage{
singleWrite(d, name = NA)
}
\arguments{
\item{d}{LiPD data}

\item{name}{Data set name}
}
\value{
none
}
\description{
Parse one data record
}
\keyword{internal}

