% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_extract.R
\name{splitInterpretationByScope}
\alias{splitInterpretationByScope}
\title{Split interpretation by scope}
\usage{
splitInterpretationByScope(TS)
}
\arguments{
\item{TS}{}
}
\value{
split TS
}
\description{
Split interpretation by scope
}
