% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{stripExtension}
\alias{stripExtension}
\title{stripExtension}
\usage{
stripExtension(filename)
}
\arguments{
\item{filename}{file name or path}
}
\value{
dataset name without extension
}
\description{
stripExtension
}
