% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{swapGeoCoordinates}
\alias{swapGeoCoordinates}
\title{An old bug caused some geo coordinates to be reversed. This will switch them back to normal.}
\usage{
swapGeoCoordinates(d)
}
\arguments{
\item{d}{Metadata}
}
\value{
d Modified metadata
}
\description{
An old bug caused some geo coordinates to be reversed. This will switch them back to normal.
}
\keyword{internal}
