% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_helpers.R
\name{tableToList}
\alias{tableToList}
\title{Convert table type: data frame to list}
\usage{
tableToList(table)
}
\arguments{
\item{table}{Data table}
}
\value{
table Converted data table
}
\description{
Convert table type: data frame to list
}
\keyword{internal}

