% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy2.R
\name{tsPluck}
\alias{tsPluck}
\title{Pull a single entry list object ot the top level (internal use)}
\usage{
tsPluck(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Pull a single entry list object ot the top level (internal use)
}
