% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unindex_geo}
\alias{unindex_geo}
\title{Convert geo from semi-flat structure back to original GeoJSON structure.}
\usage{
unindex_geo(d)
}
\arguments{
\item{d}{Metadata}
}
\value{
d Modified metadata
}
\description{
Convert geo from semi-flat structure back to original GeoJSON structure.
}
\keyword{internal}

