% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unindexGeo}
\alias{unindexGeo}
\title{Convert geo from semi-flat structure back to original GeoJSON structure.}
\usage{
unindexGeo(d)
}
\arguments{
\item{d}{Metadata}
}
\value{
d Modified metadata
}
\description{
Convert geo from semi-flat structure back to original GeoJSON structure.
}
\keyword{internal}

