% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_helpers.R
\name{unzipper}
\alias{unzipper}
\title{Unzip LiPD file to the temporary directory}
\usage{
unzipper(file, tmp)
}
\arguments{
\item{file}{One LiPD file to unzip}

\item{tmp}{Temporary directory}
}
\value{
none
}
\description{
Unzip LiPD file to the temporary directory
}
\keyword{internal}

