% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{update_lipd_v1_1}
\alias{update_lipd_v1_1}
\title{Update LiPD v1.0 to v1.1
- chronData entry is a list that allows multiple tables
- paleoData entry is a list that allows multiple tables
- chronData now allows measurement, model, summary, ensemble, calibratedAges tables
- Added 'lipdVersion' key}
\usage{
update_lipd_v1_1(d)
}
\arguments{
\item{d}{Metadata}
}
\value{
d Metadata
}
\description{
Update LiPD v1.0 to v1.1
- chronData entry is a list that allows multiple tables
- paleoData entry is a list that allows multiple tables
- chronData now allows measurement, model, summary, ensemble, calibratedAges tables
- Added 'lipdVersion' key
}
\keyword{internal}

