% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{update_lipd_v1_2_section}
\alias{update_lipd_v1_2_section}
\title{Update LiPD v1.1 to v1.2 - one section}
\usage{
update_lipd_v1_2_section(d, pc)
}
\arguments{
\item{list}{d: Metadata}

\item{char}{pc: paleoData or chronData}
}
\value{
list d: Metadata
}
\description{
Update LiPD v1.1 to v1.2 - one section
}
\keyword{internal}
