% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{update_lipd_v1_3_structure}
\alias{update_lipd_v1_3_structure}
\title{Update the structure for summary and ensemble tables}
\usage{
update_lipd_v1_3_structure(d)
}
\arguments{
\item{d}{Metadata}
}
\value{
d Metadata
}
\description{
Update the structure for summary and ensemble tables
}
\keyword{internal}
