% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{update_lipd_version}
\alias{update_lipd_version}
\title{Use the current version number to determine where to start updating from. Use "chain versioning" to make it
modular. If a file is a few versions behind, convert to EACH version until reaching current. If a file is one
version behind, it will only convert once to the newest.}
\usage{
update_lipd_version(d)
}
\arguments{
\item{d}{Metadata}
}
\value{
d Metadata
}
\description{
Use the current version number to determine where to start updating from. Use "chain versioning" to make it
modular. If a file is a few versions behind, convert to EACH version until reaching current. If a file is one
version behind, it will only convert once to the newest.
}
\keyword{internal}
