% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{verifyOutputFilename}
\alias{verifyOutputFilename}
\title{Check if output filename has invalid filename characters. Replace if necessary
R will not zip directories with certain characters.}
\usage{
verifyOutputFilename(x)
}
\arguments{
\item{x}{String}
}
\value{
x String
}
\description{
Check if output filename has invalid filename characters. Replace if necessary
R will not zip directories with certain characters.
}
\keyword{internal}

