% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvs.R
\name{write_csv_to_file}
\alias{write_csv_to_file}
\title{Write out each CSV file for this LiPD recorde
csvs format: [ some_filename.csv $columns.data ]}
\usage{
write_csv_to_file(csvs)
}
\arguments{
\item{list}{csvs: CSV data}
}
\value{
bool success: CSV write success or fail
}
\description{
Write out each CSV file for this LiPD recorde
csvs format: [ some_filename.csv $columns.data ]
}
\keyword{internal}

