% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_csv.R
\name{writeCsvs}
\alias{writeCsvs}
\title{Write out each CSV file for this LiPD record
csv.data format: [ some_filename.csv $columns.data ]}
\usage{
writeCsvs(csv.data)
}
\arguments{
\item{csv.data}{List of Lists of csv column data}
}
\value{
success Boolean for successful csv write
}
\description{
Write out each CSV file for this LiPD record
csv.data format: [ some_filename.csv $columns.data ]
}
\keyword{internal}

