% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{writeLipd}
\alias{writeLipd}
\title{Write LiPD data onto disk as LiPD files}
\usage{
writeLipd(D, path)
}
\arguments{
\item{list}{D: LiPD data}

\item{char}{path: Destination path}
}
\value{
none:
}
\description{
Write LiPD data onto disk as LiPD files
}
\keyword{internal}

