% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lipds_file.R
\name{writeLipdFile}
\alias{writeLipdFile}
\title{Main function. Run all write sub-routines for one LiPD record}
\usage{
writeLipdFile(d, name)
}
\arguments{
\item{d}{Metadata}

\item{name}{Name of current LiPD record}
}
\value{
none
}
\description{
Main function. Run all write sub-routines for one LiPD record
}
\keyword{internal}

