% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lipds_main.R
\name{writeLipds}
\alias{writeLipds}
\title{Main function. Loop through all LiPD records, saving, and then cleaning up}
\usage{
writeLipds(D)
}
\arguments{
\item{D}{LiPD Library}
}
\value{
none
}
\description{
Main function. Loop through all LiPD records, saving, and then cleaning up
}

