% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zips.R
\name{zipper}
\alias{zipper}
\title{Zip a directory, and move up a level}
\usage{
zipper(dir_original, dir_tgt, dir_tmp, bag.success, dsn)
}
\arguments{
\item{char}{dir_tgt: Directory to be zipped}

\item{char}{dir_tmp  Directory that holds resulting zip file}
}
\value{
none
}
\description{
Zip a directory, and move up a level
}
\keyword{internal}

