% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_helpers.R
\name{bagit}
\alias{bagit}
\title{Since we don't have a way of getting the bagit module in R,
all we can do is use the default bag function by calling the
full python file on a directory. This will create a bag.}
\usage{
bagit(data.dir, initial.dir)
}
\arguments{
\item{path}{The path to the directory that needs to be bagged}
}
\value{
none
}
\description{
Since we don't have a way of getting the bagit module in R,
all we can do is use the default bag function by calling the
full python file on a directory. This will create a bag.
}
\keyword{internal}

