% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{collapse_block}
\alias{collapse_block}
\title{Collapse all blocks listed: funding, publication calibration, interpretation
These follow the regex format of "<key1><idx>_<key2>"}
\usage{
collapse_block(entry, l, key)
}
\arguments{
\item{list}{entry: Time series entry}

\item{list}{l: Metadata (to append to)}

\item{char}{key: Key from time series entry}
}
\value{
list l: Metadata
}
\description{
Collapse all blocks listed: funding, publication calibration, interpretation
These follow the regex format of "<key1><idx>_<key2>"
}

