% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{collapse_pc}
\alias{collapse_pc}
\title{Collapse time series section; paleo or chron}
\usage{
collapse_pc(d, entry, pc)
}
\arguments{
\item{list}{d: Metadata}

\item{list}{entry: Time series entry}

\item{char}{pc: paleoData or chronData}
}
\value{
list d: Metadata
}
\description{
Collapse time series section; paleo or chron
}

