% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lipds_csv.R
\name{collectCsvs}
\alias{collectCsvs}
\title{Collect csvs main. Collect and remove csv "values" fields.}
\usage{
collectCsvs(dsn, d)
}
\arguments{
\item{dsn}{Dataset name}

\item{d}{Metadata}
}
\value{
all.data Final split of metadata and csv data
}
\description{
Collect csvs main. Collect and remove csv "values" fields.
}
\keyword{internal}

