% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_lipds_csv.R
\name{collectCsvSection}
\alias{collectCsvSection}
\title{Collect and remove csv from one section: Paleo or chron
csv.data format: [ some_filename.csv $columns.data ]}
\usage{
collectCsvSection(all.data, keys, name)
}
\arguments{
\item{keys}{Section keys}

\item{name}{LiPD data set name}

\item{d}{Metadata w. values}

\item{csv.data}{Running collection of csv data}
}
\value{
all.data List holding the running collection of separated csv and metadata
}
\description{
Collect and remove csv from one section: Paleo or chron
csv.data format: [ some_filename.csv $columns.data ]
}
\keyword{internal}

