% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{filterTs}
\alias{filterTs}
\title{Find all the time series entries that match a given search expression, 
and return a new time series with the matching entries
new.ts = filterTs(ts, "archiveType == marine sediment")
Valid operators : ==, =, <=, >=, <, >}
\usage{
filterTs(ts, expression)
}
\arguments{
\item{ts}{Time series}

\item{expression}{Search expression}
}
\value{
new.ts Time series
}
\description{
Find all the time series entries that match a given search expression, 
and return a new time series with the matching entries
new.ts = filterTs(ts, "archiveType == marine sediment")
Valid operators : ==, =, <=, >=, <, >
}

