% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{filterTs}
\alias{filterTs}
\title{Find all the time series objects that match a given search expression, 
and return a new time series with the matching objects}
\usage{
filterTs(ts, expression)
}
\arguments{
\item{ts}{Time series}

\item{expression}{Search expression}
}
\value{
new.ts : Time series
}
\description{
Valid operators : ==, =, <=, >=, <, >
}
\examples{

# Time series
ts = [ object1, object2, object3, object4 ]

# Example 1
new.ts = filterTs(ts, "archiveType == marine sediment")
# result 
[object1, object3, object4]

# Example 2
new.ts = filterTs(ts, "paleoData_variableName == d18O")
# result
[object2]


}
\author{
Chris Heiser
}

