% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_matches}
\alias{get_matches}
\title{Use the regex match groups and the time series to compile two lists: matching indices, and matching entries.}
\usage{
get_matches(ts, m, exact)
}
\arguments{
\item{ts}{Time series}

\item{m}{Regex match groups}
}
\description{
Use the regex match groups and the time series to compile two lists: matching indices, and matching entries.
}
\author{
Chris Heiser
}
\keyword{internal}
