% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{get_table}
\alias{get_table}
\title{Get the target table}
\usage{
get_table(d, current, pc)
}
\arguments{
\item{list}{d: Metadata}

\item{list}{current: Current time series entry}

\item{char}{pc: paleoData or chronData}
}
\value{
list table: Metadata
}
\description{
Get the target table
}
