% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_idx_to_name.R
\name{indexSectionRead}
\alias{indexSectionRead}
\title{Change index-by-number for one section}
\usage{
indexSectionRead(d, keys)
}
\arguments{
\item{d}{LiPD metadata}

\item{keys}{Section keys}
}
\value{
d Modified LiPD metadata
}
\description{
Change index-by-number for one section
}
\keyword{internal}

