% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.R
\name{lipd_read}
\alias{lipd_read}
\title{Loads a LiPD file from local path. Unzip, read, and process data
Steps: create tmp, unzip lipd, read files into memory, manipulate data, move to original dir, delete tmp.}
\usage{
lipd_read(path)
}
\arguments{
\item{lpds_ext}{List of LiPD filenames to read (w/ .lpd extension)}

\item{working.dir}{Directory that contains the target files}
}
\value{
d LiPD File
}
\description{
Loads a LiPD file from local path. Unzip, read, and process data
Steps: create tmp, unzip lipd, read files into memory, manipulate data, move to original dir, delete tmp.
}
\keyword{internal}

