% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_merge.R
\name{merge_csv_columns}
\alias{merge_csv_columns}
\title{Merge values into each column}
\usage{
merge_csv_columns(csvs, meta)
}
\arguments{
\item{list}{meta: Table metadata, sorted by column}
}
\value{
list meta: Table metadata
}
\description{
Merge values into each column
}
\keyword{internal}
