% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_merge.R
\name{merge_csv_model}
\alias{merge_csv_model}
\title{Add csv data to each column in chron model}
\usage{
merge_csv_model(models, pc, crumbs)
}
\arguments{
\item{list}{models: Metadata}

\item{char}{pc: paleoData or chronData}

\item{char}{crumbs: Hierarchy crumbs}
}
\value{
list models: Metadata
}
\description{
Add csv data to each column in chron model
}
\keyword{internal}

