% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_merge.R
\name{merge_csv_section}
\alias{merge_csv_section}
\title{Merge CSV into each section}
\usage{
merge_csv_section(section, crumbs, csvs)
}
\arguments{
\item{list}{section: Metadata}

\item{char}{crumbs: paleo or chron}

\item{list}{csvs: CSV, sorted by filename}
}
\value{
list section: Metadata
}
\description{
Merge CSV into each section
}
\keyword{internal}

