% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{put_base_data}
\alias{put_base_data}
\title{Put in paleoData and chronData as the base for this dataset using the oroginal dataset data.}
\usage{
put_base_data(entry, raw_datasets, dsn, force, mode)
}
\arguments{
\item{list}{d: Metadata}

\item{char}{pc: paleoData or chronData}

\item{bool}{force: Build dataset without original data from global}

\item{char}{mode: paleo or chron mode}
}
\value{
list d: Metadata
}
\description{
Put in paleoData and chronData as the base for this dataset using the oroginal dataset data.
}
