% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{readLipd}
\alias{readLipd}
\title{Read LiPD files into R workspace}
\usage{
readLipd(path = NULL)
}
\arguments{
\item{path}{Source path (optional)}
}
\value{
D : LiPD dataset(s)
}
\description{
Read LiPD files into R workspace
}
\examples{
read in multiple datasets - no path argument
D <- readLipd()   # choose option 'd' for directory

read in multiple datasets - with path argument
D <- readLipd("/Users/bobsmith/Desktop/lipd_files")

read in one dataset - no path argument
L <- readLipd()  # choose option "s" for single file

read in one dataset - with path argument
L <- readLipd("/Users/bobsmith/Desktop/lipd_files/dataset.lpd")

}
\author{
Chris Heiser
}
\keyword{internal}

