% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_main.R
\name{readLipd}
\alias{readLipd}
\title{Main LiPD reading function. Combines all processes into one.}
\usage{
readLipd()
}
\value{
D LiPD Library
}
\description{
Main LiPD reading function. Combines all processes into one.
}
\keyword{internal}

