% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipds_helpers.R
\name{stripExtension}
\alias{stripExtension}
\title{Remove the file extension from string name}
\usage{
stripExtension(filename)
}
\arguments{
\item{filename}{One LiPD filename}
}
\value{
x LiPD filename w/o extension
}
\description{
Remove the file extension from string name
}
\keyword{internal}

