% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvs.R
\name{write_csv_to_file}
\alias{write_csv_to_file}
\title{Write out each CSV file for this LiPD record
csv.data format: [ some_filename.csv $columns.data ]}
\usage{
write_csv_to_file(csv.data)
}
\arguments{
\item{csv.data}{List of Lists of csv column data}
}
\value{
success Boolean for successful csv write
}
\description{
Write out each CSV file for this LiPD record
csv.data format: [ some_filename.csv $columns.data ]
}
\keyword{internal}

