% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{writeLipd}
\alias{writeLipd}
\title{Write LiPD data onto disk as LiPD files}
\usage{
writeLipd(D, path)
}
\arguments{
\item{D}{LiPD datasets : list}

\item{path}{Destination path : char}
}
\value{
none
}
\description{
Write LiPD data onto disk as LiPD files
}
\examples{

# write - without path argument
writeLipd(D)

# write - with path argument
writeLipd(D, "/Users/bobsmith/Desktop/lipd_files")

}
\author{
Chris Heiser
}
\keyword{internal}

