% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zips.R
\name{zipper}
\alias{zipper}
\title{Zip a directory, and move up a level}
\usage{
zipper(dir_original, dir_tmp, dsn)
}
\arguments{
\item{char}{dir_original: Directory to move the LiPD to}

\item{char}{dir_tmp:  Temp directory}
}
\value{
none
}
\description{
Zip a directory, and move up a level
}
\keyword{internal}

