% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_helpers.R
\name{zipper}
\alias{zipper}
\title{Zip a directory, and move up a level}
\usage{
zipper(dir, tmp)
}
\arguments{
\item{dir}{Directory to be zipped}

\item{tmp}{Directory that holds resulting zip file}
}
\value{
none
}
\description{
Zip a directory, and move up a level
}
\keyword{internal}

